pyproject.toml
[tool.poetry]
name = "rand"
version = "0.1.0"
description = "Generate String from regex pattern"
authors = ["kororo"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/kororo/rand"
keywords = ["random", "generator", "string", "regex"]
packages = [
    { include = "rand" }
]
[tool.poetry.dependencies]
python = "^3.6"
pytest-cov = { version = "^2.0" }

[tool.poetry.extras]
test = ["pytest"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
